CAR_RENTAL_PROJECT
This Car Rental System project is designed to aid the car rental company to enable renting of cars through an online system. It helps the users to search for available cars view profile and book the cars for the time period. It has a user-friendly interface which helps the user to check for cars and rent them for the period specified. They could also make payment online. The rental cars shall be categorized into economy, premium etc. Based on the type of car required by the customer, the user shall be able to make bookings. The use of internet technology has made it easy for the customers to rent a car any time. This Car Rental System makes the bookings easy. It saves time and labor. The tool shall ask the user for information such as the date and time of journey, type of car etc. Also, it will need an identification number. Using these details, the tool shall help the customer to book a car for the journey.